+++ Enigma/Enigma.Frontend/ViewModels/PreNatalAspectSelectionViewModel.cs
@@ -1,0 +1,50 @@
+// Enigma Astrology Research.
+// Copyright (c) 2025 Jan Kampherbeek.
+// Enigma is open source.
+// Please check the file copyright.txt in the root of the source for further details.
+
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.Linq;
+using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.Input;
+using CommunityToolkit.Mvvm.Messaging;
+using Enigma.Domain.References;
+using Enigma.Frontend.Ui.Messaging;
+using Enigma.Frontend.Ui.Models;
+using Enigma.Frontend.Ui.State;
+using Enigma.Frontend.Ui.Support;
+using Enigma.Frontend.Ui.WindowsFlow;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Enigma.Frontend.Ui.ViewModels;
+
+public partial class PreNatalAspectSelectionViewModel: ObservableObject
+{
+      
+    private const string VM_IDENTIFICATION = ChartsWindowsFlow.PRENATAL_ASPECT_SELECTION;
+    
+    [ObservableProperty] private ObservableCollection<SelectableAspectDetails> _allAspectDetails;
+
+    public PreNatalAspectSelectionViewModel()
+    {
+        var model = App.ServiceProvider.GetRequiredService<PreNatalAspectSelectionModel>();
+        _allAspectDetails =  new ObservableCollection<SelectableAspectDetails>(model.GetAllAspectDetails());
+    }
+  
+    [RelayCommand]
+    private void Continue()
+    {
+        List<AspectTypes> selectedAspects = (from item in _allAspectDetails 
+            where item.Selected select item.Aspect).ToList();
+        DataVaultCharts.Instance.CurrentAspectsSelection = selectedAspects;
+        WeakReferenceMessenger.Default.Send(new CloseMessage(VM_IDENTIFICATION));
+    }
+
+    
+    [RelayCommand]
+    private static void Help()
+    {
+        WeakReferenceMessenger.Default.Send(new HelpMessage(VM_IDENTIFICATION));
+    }
+}