+++ Enigma/Enigma.Frontend/ViewModels/PrenatalViewModel.cs
@@ -4,12 +4,16 @@
 // Please check the file copyright.txt in the root of the source for further details.
 
 using System.Collections.Generic;
+using System.Windows;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using CommunityToolkit.Mvvm.Messaging;
+using Enigma.Domain.Dtos;
+using Enigma.Domain.References;
 using Enigma.Frontend.Ui.Messaging;
 using Enigma.Frontend.Ui.Models;
 using Enigma.Frontend.Ui.PresentationFactories;
+using Enigma.Frontend.Ui.State;
 using Enigma.Frontend.Ui.WindowsFlow;
 using Microsoft.Extensions.DependencyInjection;
 using Serilog;
@@ -30,6 +34,8 @@ public partial class PrenatalViewModel : ObservableObject
     [ObservableProperty] private bool _useRetrogradeDirect;
     [ObservableProperty] private List<PresentablePreNatalMoment> _preNatalMoments = new();
     [ObservableProperty] private List<PresentablePreNatalEvent> _preNatalEvents = new();
+    [ObservableProperty] private string _selectedFactors;
+    [ObservableProperty] private string _selectedAspects;
     
     [RelayCommand]
     private void Help()
@@ -51,10 +57,67 @@ public partial class PrenatalViewModel : ObservableObject
         UseEclipses = true;
         UseIngresses = false;
         UseRetrogradeDirect = false;
-        
-      
+        var selectedFactors = new List<ChartPoints>
+        {
+            ChartPoints.Sun,
+            ChartPoints.Mercury,
+            ChartPoints.Venus,
+            ChartPoints.Mars,
+            ChartPoints.Jupiter,
+            ChartPoints.Saturn,
+            ChartPoints.Uranus,
+            ChartPoints.Neptune,
+            ChartPoints.Pluto
+        };
+        var selectedAspects = new List<AspectTypes>
+        {
+            AspectTypes.Conjunction
+        };
+        _preNatalModel.selectedFactors = selectedFactors;
+        DataVaultCharts.Instance.CurrentPointsSelection = selectedFactors;
+        _preNatalModel.selectedAspects = selectedAspects;
+        DataVaultCharts.Instance.CurrentAspectsSelection = selectedAspects;
     }
 
+    private void DefineFactors()
+    {
+        WeakReferenceMessenger.Default.Send(new OpenMessage(VM_IDENTIFICATION,
+            ChartsWindowsFlow.PRENATAL_FACTOR_SELECTION));
+        var selection = DataVaultCharts.Instance.CurrentPointsSelection;
+        if (selection == null) return;
+        if (selection.Count < 1)
+        {
+            MessageBox.Show("Please select at least one point." );
+        }
+        else
+        {
+            _preNatalModel.selectedFactors = selection;
+            DataVaultCharts.Instance.CurrentPointsSelection = selection;
+            SelectedFactors = _preNatalModel.GetSelectedFactorsGlyphs();
+            Populate();       
+        }
+    }
+    
+    private void DefineAspects()
+    {
+        WeakReferenceMessenger.Default.Send(new OpenMessage(VM_IDENTIFICATION,
+            ChartsWindowsFlow.PRENATAL_ASPECT_SELECTION));
+        var selection = DataVaultCharts.Instance.CurrentAspectsSelection;
+        if (selection == null) return;
+        if (selection.Count < 1)
+        {
+            MessageBox.Show("Please select at least one aspect." );
+        }
+        else
+        {
+            _preNatalModel.selectedAspects = selection;
+            DataVaultCharts.Instance.CurrentAspectsSelection = selection;
+            SelectedAspects = _preNatalModel.GetSelectedAspectsGlyphs();
+            Populate();       
+        }
+    }
+    
+    
     public void Populate()
     {
         _preNatalModel.useAspects = UseAspects;
@@ -62,31 +125,44 @@ public partial class PrenatalViewModel : ObservableObject
         _preNatalModel.useIngresses = UseIngresses;
         _preNatalModel.useRetrogradeDirect = UseRetrogradeDirect;
         PreNatalMoments = _preNatalModel.GetPrenatalMoments();
-        PreNatalEvents = _preNatalModel.GetPrenatalEvents();              
+        PreNatalEvents = _preNatalModel.GetPrenatalEvents();  
+        SelectedFactors = _preNatalModel.GetSelectedFactorsGlyphs();
+        SelectedAspects = _preNatalModel.GetSelectedAspectsGlyphs();
     }
     
     public void UpdateAspects(bool useIt)
     {
-        _useAspects = useIt;
+        UseAspects = useIt;
         Populate();
     }
     
     public void UpdateEclipses(bool useIt)
     {
-        _useEclipses = useIt;
+        UseEclipses = useIt;
         Populate();
     }
     
     public void UpdateRetrogradeDirect(bool useIt)
     {
-        _useRetrogradeDirect = useIt;
+        UseRetrogradeDirect = useIt;
         Populate();
     }
     
     public void UpdateIngresses(bool useIt)
     {
-        _useIngresses = useIt;
+        UseIngresses = useIt;
         Populate();
     }
+
+    [RelayCommand]
+    public void ChangePoints()
+    {
+        DefineFactors();       
+    }
     
+    [RelayCommand]
+    public void ChangeAspects()
+    {
+        DefineAspects();       
+    }
 }