+++ Enigma/Enigma.Frontend/Views/PreNatalAspectSelectionWindow.xaml.cs
@@ -1,0 +1,16 @@
+// Enigma Astrology Research.
+// Copyright (c) 2025 Jan Kampherbeek.
+// Enigma is open source.
+// Please check the file copyright.txt in the root of the source for further details.
+
+using System.Windows;
+
+namespace Enigma.Frontend.Ui.Views;
+
+public partial class PreNatalAspectSelectionWindow : Window
+{
+    public PreNatalAspectSelectionWindow()
+    {
+        InitializeComponent();
+    }
+}